Changed primitive double to int

+++ project/pa-app/src/main/java/dk/dtu/pa/geometry/ResizeShapes.java
@@ -1,5 +1,5 @@
 package dk.dtu.pa.geometry;
 
 public interface ResizeShapes {
-    void resize(double base, double high);
+    void resize(int base, int high);
 }
