+++ project/pa-app/src/main/java/dk/dtu/pa/geometry/Triangle.java
@@ -5,17 +5,17 @@ package dk.dtu.pa.geometry;
 public class Triangle extends Shapes {
     private final int side = 3;
 
-    public Triangle(Type type, double base, double high) {
+    public Triangle(Type type, int base, int high) {
         this.type = type;
         this.area = computeArea(base, high);
         this.perimeter = computePerimeter(base);
     }
 
-    public double computeArea(double base, double high) {
+    public int computeArea(int base, int high) {
         return (base * high) / 2;
     }
 
-    public double computePerimeter(double base) {
+    public int computePerimeter(int base) {
         return base * 3;
     }
 
@@ -24,11 +24,11 @@ public class Triangle extends Shapes {
         return this.side;
     }
 
-    public double getArea(){
+    public int getArea(){
         return this.area;
     }
 
-    public double getPerimeter() {
+    public int getPerimeter() {
         return this.perimeter;
     }
